reorder fns
